requirements.txt
async-timeout==4.0.3
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
lmdb==1.4.1
logging-tree==1.9
MarkupSafe==2.1.3
pyfakeuse==0.0.8
pylogconf==0.0.37
pytconf==0.0.98
python-decouple==3.8
pyunique==0.0.10
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
spotipy==2.23.0
termcolor==2.3.0
tqdm==4.66.1
urllib3==2.1.0
Werkzeug==3.0.1
yattag==1.15.2
zipp==3.17.0
